Add files via upload
In this project first you can first register yourself as a student, volunteer or as a organizer.
Then when you login as student you can view certain events, register to the certain events or even deregister for the certain event.
When you login as a admin you can add the events that is set the event name, date,venue. you add the organizers where the organizes can organize certain events.
The volunteer can voulunteer the events. Here we have a special of report generation where you can give the start date and the end date where you can  see the events held in the dates that is mentioned.
This is a college event management system. There u can find two distinct types that is the react app and the spring app.
First when u open the app in the visual studio .
Go to application.properties, you will have to change the password, and just check what the username is if it is not root then change that also.
	open the springapp folder
		right click…select “open in terminal”
	open the reactapp folder
		right click…select “open in terminal”
open mysql
	create database test;
	use test;
In springapp terminal :mvn spring-boot:run
In reactapp terminal : npm start